fix(ProcessSample): Fix unique contraint requirement

+++ src/ProcessSample.php
@@ -14,13 +14,15 @@ final class ProcessSample
         $samples = Sample::groupBy(['term', 'type', 'gender', 'canonical'])->selectRaw('term, type, gender, canonical, count(*) as total')->get();
         foreach ($samples as $sample) {
             if ($sample->total > 1) {
-                $term = new Term();
-                $term->term = $sample->term;
-                $term->type = $sample->type;
-                $term->gender = $sample->gender;
-                $term->canonical = $sample->canonical;
-                if ($term->save()) {
-                    Sample::where('term', $sample->term)->where('type', $sample->type)->where('gender', $sample->gender)->where('canonical', $sample->canonical)->delete();
+                if (!Term::where('term', $sample->term)->where('type', $sample->type)->exists()) {
+                    $term = new Term();
+                    $term->term = $sample->term;
+                    $term->type = $sample->type;
+                    $term->gender = $sample->gender;
+                    $term->canonical = $sample->canonical;
+                    if ($term->save()) {
+                        Sample::where('term', $sample->term)->where('type', $sample->type)->where('gender', $sample->gender)->where('canonical', $sample->canonical)->delete();
+                    }
                 }
             }
         }
